requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
torch==2.2.2
torchvision==0.17.2
Pillow==10.4.0
python-multipart==0.0.12
